# cmd/api/docs/swagger.yaml
definitions:
...
  schemas.Response:
    properties:
      body: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
  schemas.Supplier:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  schemas.SupplierCreate:
    properties:
      address:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - name
    type: object
  schemas.SupplierUpdate:
    properties:
      address:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - id
    - name
    type: object
...
info:
  contact: {}
  description: This is a api to app noa gestion
  termsOfService: http://swagger.io/terms/
  title: APP NOA Gestion
  version: "1.0"
paths:
...
  /api/v1/supplier/{id}:
    get:
      consumes:
      - application/json
      description: Get a supplier by its ID within a specified workplace.
      parameters:
      - description: ID of the supplier
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Supplier obtained with success
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  $ref: '#/definitions/schemas.Supplier'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Supplier not found
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Get Supplier By ID
      tags:
      - Supplier
  /api/v1/supplier/create:
    post:
      consumes:
      - application/json
      description: Creates a new supplier within the specified workplace.
      parameters:
      - description: Details of the supplier to create
        in: body
        name: supplier
        required: true
        schema:
          $ref: '#/definitions/schemas.SupplierCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Supplier created successfully
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Model is invalid
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Create Supplier
      tags:
      - Supplier
  /api/v1/supplier/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a supplier based on the provided ID and workplace context.
      parameters:
      - description: ID of the supplier
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Supplier deleted with success
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Supplier not found
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Delete Supplier
      tags:
      - Supplier
  /api/v1/supplier/get_all:
    get:
      consumes:
      - application/json
      description: Get All Suppliers
      produces:
      - application/json
      responses:
        "200":
          description: Suppliers obtained with success
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/schemas.Supplier'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Get All Suppliers
      tags:
      - Supplier
  /api/v1/supplier/get_by_name:
    get:
      consumes:
      - application/json
      description: Fetches suppliers from either laundry or workshop based on the
        provided name and workplace.
      parameters:
      - description: Name of the Supplier
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of suppliers
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/schemas.Supplier'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Get Supplier By Name
      tags:
      - Supplier
  /api/v1/supplier/update:
    put:
      consumes:
      - application/json
      description: Update a supplier's information from the specified workplace.
      parameters:
      - description: Supplier information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schemas.SupplierUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Supplier updated with success
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Auth is required
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Not Authorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Supplier not found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Model is invalid
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Update Supplier
      tags:
      - Supplier
...
securityDefinitions:
  CookieAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"